# [PERF] stock,mrp: optimize _search_product_quantity