The performance when searching by `qty_available`, `virtual_available`, `free_qty`, `incoming_qty`, or `outgoing_qty` on databases with a large number of `product.product`s is very poor. We improve the performance by moving `_compute_quantities_dict()` up the call stack in `_search_product_quantity()`.

Previously, `_search_product_quantity()` would compute the fields for each product individually in a loop, generating N queries:

Instead, we directly call `_compute_quantities_dict()` on the recordset, which will perform a single read on quants/moves and compute the quantities for all products at once. This results in a constant number of queries relative to the number of products in the database. Some additional optimizations are made in `_compute_quantities_dict()` to avoid additional queries.

Benchmark

| `product.product` count | Before                 | After                 |
| ----------------------- | ---------------------- | --------------------- |
| 5,000                   | 1.48s, 62 queries      | 0.31s, 38 queries     |
| 50,000                  | 13.44s, 351 queries    | 2.00s, 36 queries     |
| 500,000                 | 201.36s, 4,554 queries | 29.89s, 28 queries    |

Average 6x speedup

opw-4930856