@robodoo 2025-08-19
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/223471.png)](https://mergebot.odoo.com/odoo/odoo/pull/223471)